requirements.txt
aiohttp==3.8.1
APScheduler==3.9.1
fastapi==0.75.2
loguru==0.6.0
PyYAML==6.0
sqlmodel==0.0.6
starlette==0.17.1
